move of BibliographyDB
