+++ bibliography/static/js/es6_modules/bibliography/database.js
