+++ bibliography/static/js/exporter.es6.js
@@ -1,6 +1,6 @@
 import {BibLatexExporter} from "./es6_modules/bibliography/exporter/biblatex"
 import {BibLatexImporter} from "./es6_modules/bibliography/importer/biblatex"
-import {BibliographyDB} from "./es6_modules/bibliography/bibliographyDB"
+import {BibliographyDB} from "./es6_modules/bibliography/database"
 
 window.BibLatexExporter = BibLatexExporter
 window.BibLatexImporter = BibLatexImporter
