+++ document/static/js/document-overview.es5.js
@@ -2011,7 +2011,7 @@ var _native = require("./native");
 
 var _native2 = require("../importer/native");
 
-var _bibliographyDB = require("../bibliography/bibliographyDB");
+var _database = require("../bibliography/database");
 
 var afterCopy = function afterCopy(noErrors, returnValue, editor, callback) {
     $.deactivateWait();
@@ -2056,7 +2056,7 @@ var savecopy = exports.savecopy = function savecopy(aDocument, editor, user, cal
             importAsUser(aDocument, shrunkImageDB, shrunkBibDB, images, editor, user, callback);
         });
     } else {
-        var bibGetter = new _bibliographyDB.BibliographyDB(aDocument.owner.id, false, false, false);
+        var bibGetter = new _database.BibliographyDB(aDocument.owner.id, false, false, false);
         bibGetter.getBibDB(function (bibDB, bibCats) {
             usermediaHelpers.getAnImageDB(aDocument.owner.id, function (anImageDB) {
                 (0, _native.exportNative)(aDocument, anImageDB, bibDB, function (aDocument, shrunkImageDB, shrunkBibDB, images) {
@@ -2067,7 +2067,7 @@ var savecopy = exports.savecopy = function savecopy(aDocument, editor, user, cal
     }
 };
 
-},{"../bibliography/bibliographyDB":2,"../importer/native":30,"./native":24}],17:[function(require,module,exports){
+},{"../bibliography/database":2,"../importer/native":30,"./native":24}],17:[function(require,module,exports){
 "use strict";
 
 Object.defineProperty(exports, "__esModule", {
@@ -2582,7 +2582,7 @@ var _zip = require("./zip");
 
 var _htmlTemplates = require("./html-templates");
 
-var _bibliographyDB = require("../bibliography/bibliographyDB");
+var _database = require("../bibliography/database");
 
 var _base = require("./base");
 
@@ -2713,7 +2713,7 @@ var HTMLExporter = exports.HTMLExporter = (function (_BaseHTMLExporter) {
             _this2.exportOne();
         } else {
             (function () {
-                var bibGetter = new _bibliographyDB.BibliographyDB(doc.owner.id, false, false, false);
+                var bibGetter = new _database.BibliographyDB(doc.owner.id, false, false, false);
                 bibGetter.getBibDB(function () {
                     that.bibDB = bibGetter.bibDB;
                     that.exportOne();
@@ -2799,7 +2799,7 @@ var HTMLExporter = exports.HTMLExporter = (function (_BaseHTMLExporter) {
     return HTMLExporter;
 })(BaseHTMLExporter);
 
-},{"../bibliography/bibliographyDB":2,"../citations/format":5,"./base":15,"./html-templates":20,"./json":22,"./tools":25,"./zip":28,"katex":32}],22:[function(require,module,exports){
+},{"../bibliography/database":2,"../citations/format":5,"./base":15,"./html-templates":20,"./json":22,"./tools":25,"./zip":28,"katex":32}],22:[function(require,module,exports){
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
@@ -2896,7 +2896,7 @@ var _base = require("./base");
 
 var _biblatex = require("../bibliography/exporter/biblatex");
 
-var _bibliographyDB = require("../bibliography/bibliographyDB");
+var _database = require("../bibliography/database");
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
@@ -3280,7 +3280,7 @@ var LatexExporter = exports.LatexExporter = (function (_BaseLatexExporter) {
             _this2.exportOne();
         } else {
             (function () {
-                var bibGetter = new _bibliographyDB.BibliographyDB(doc.owner.id, false, false, false);
+                var bibGetter = new _database.BibliographyDB(doc.owner.id, false, false, false);
                 bibGetter.getBibDB(function () {
                     that.bibDB = bibGetter.bibDB;
                     that.exportOne();
@@ -3328,7 +3328,7 @@ var LatexExporter = exports.LatexExporter = (function (_BaseLatexExporter) {
     return LatexExporter;
 })(BaseLatexExporter);
 
-},{"../bibliography/bibliographyDB":2,"../bibliography/exporter/biblatex":3,"./base":15,"./json":22,"./tools":25,"./zip":28}],24:[function(require,module,exports){
+},{"../bibliography/database":2,"../bibliography/exporter/biblatex":3,"./base":15,"./json":22,"./tools":25,"./zip":28}],24:[function(require,module,exports){
 "use strict";
 
 Object.defineProperty(exports, "__esModule", {
@@ -3342,7 +3342,7 @@ var _tools = require("./tools");
 
 var _zip = require("./zip");
 
-var _bibliographyDB = require("../bibliography/bibliographyDB");
+var _database = require("../bibliography/database");
 
 /** The current Fidus Writer filetype version.
  * The importer will not import from a different version and the exporter
@@ -3369,12 +3369,12 @@ var downloadNative = exports.downloadNative = function downloadNative(aDocument,
             exportNative(aDocument, imageDB, bibDB, exportNativeFile);
         });
     } else if (imageDB) {
-        var bibGetter = new _bibliographyDB.BibliographyDB(aDocument.owner.id, false, false, false);
+        var bibGetter = new _database.BibliographyDB(aDocument.owner.id, false, false, false);
         bibGetter.getBibDB(function (bibDB, bibCats) {
             exportNative(aDocument, imageDB, bibDB, exportNativeFile);
         });
     } else {
-        var bibGetter = new _bibliographyDB.BibliographyDB(aDocument.owner.id, false, false, false);
+        var bibGetter = new _database.BibliographyDB(aDocument.owner.id, false, false, false);
         bibGetter.getBibDB(function (bibDB, bibCats) {
             usermediaHelpers.getAnImageDB(aDocument.owner.id, function (imageDB) {
                 exportNative(aDocument, imageDB, bibDB, exportNativeFile);
@@ -3442,7 +3442,7 @@ var exportNativeFile = function exportNativeFile(aDocument, shrunkImageDB, shrun
     (0, _zip.zipFileCreator)(outputList, httpOutputList, (0, _tools.createSlug)(aDocument.title) + '.fidus', 'application/fidus+zip', false, upload, editor);
 };
 
-},{"../bibliography/bibliographyDB":2,"./json":22,"./tools":25,"./zip":28}],25:[function(require,module,exports){
+},{"../bibliography/database":2,"./json":22,"./tools":25,"./zip":28}],25:[function(require,module,exports){
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
